MAGECLOUD-3258: Cache Warm-up, Using RegEx to Warm-up Multiple Pages
