+++ src/Process/PostDeploy/WarmUp.php
@@ -6,14 +6,11 @@
 namespace Magento\MagentoCloud\Process\PostDeploy;
 
 use GuzzleHttp\Exception\RequestException;
-use Magento\MagentoCloud\Config\Stage\PostDeployInterface;
 use Magento\MagentoCloud\Http\ClientFactory;
 use Magento\MagentoCloud\Http\RequestFactory;
+use Magento\MagentoCloud\Process\PostDeploy\WarmUp\Urls;
 use Magento\MagentoCloud\Process\ProcessException;
 use Magento\MagentoCloud\Process\ProcessInterface;
-use Magento\MagentoCloud\Shell\ShellException;
-use Magento\MagentoCloud\Shell\ShellInterface;
-use Magento\MagentoCloud\Util\UrlManager;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -21,14 +18,6 @@ use Psr\Log\LoggerInterface;
  */
 class WarmUp implements ProcessInterface
 {
-    const ENTITY_CATEGORY = 'category';
-    const ENTITY_CMS_PAGE = 'cms-page';
-
-    /**
-     * @var PostDeployInterface
-     */
-    private $postDeploy;
-
     /**
      * @var ClientFactory
      */
@@ -40,49 +29,31 @@ class WarmUp implements ProcessInterface
     private $requestFactory;
 
     /**
-     * @var UrlManager
-     */
-    private $urlManager;
-
-    /**
      * @var LoggerInterface
      */
     private $logger;
 
     /**
-     * Variable for caching base urls hosts from config table
-     *
-     * @var string
+     * @var Urls
      */
-    private $baseHosts;
+    private $urls;
 
     /**
-     * @var ShellInterface
-     */
-    private $shell;
-
-    /**
-     * @param PostDeployInterface $postDeploy
      * @param ClientFactory $clientFactory
      * @param RequestFactory $requestFactory
-     * @param UrlManager $urlManager
      * @param LoggerInterface $logger
-     * @param ShellInterface $shell
+     * @param Urls $urls
      */
     public function __construct(
-        PostDeployInterface $postDeploy,
         ClientFactory $clientFactory,
         RequestFactory $requestFactory,
-        UrlManager $urlManager,
         LoggerInterface $logger,
-        ShellInterface $shell
+        Urls $urls
     ) {
-        $this->postDeploy = $postDeploy;
         $this->clientFactory = $clientFactory;
         $this->requestFactory = $requestFactory;
-        $this->urlManager = $urlManager;
         $this->logger = $logger;
-        $this->shell = $shell;
+        $this->urls = $urls;
     }
 
     /**
@@ -98,7 +69,7 @@ class WarmUp implements ProcessInterface
         try {
             $this->logger->info('Starting page warming up');
 
-            foreach ($this->getUrlsForWarmUp() as $url) {
+            foreach ($this->urls->getAll() as $url) {
                 $request = $this->requestFactory->create('GET', $url);
 
                 $promises[] = $client->sendAsync($request)->then(function () use ($url) {
@@ -126,108 +97,5 @@ class WarmUp implements ProcessInterface
         } catch (\Throwable $exception) {
             throw new ProcessException($exception->getMessage(), $exception->getCode(), $exception);
         }
-    }
-
-    /**
-     * Returns list of URLs which should be warm up.
-     *
-     * @return array
-     */
-    private function getUrlsForWarmUp(): array
-    {
-        $pages = $this->postDeploy->get(PostDeployInterface::VAR_WARM_UP_PAGES);
-        $baseUrl = rtrim($this->urlManager->getBaseUrl(), '/');
-        $urls = [];
-
-        $urlPattern = sprintf(
-            '/^(%s|%s):(\d+|\*):.{1,}/',
-            self::ENTITY_CATEGORY,
-            self::ENTITY_CMS_PAGE
-        );
-
-        foreach ($pages as $page) {
-            if (preg_match($urlPattern, $page)) {
-                $patternUrls = $this->getUrlsByPattern($page);
-                $this->logger->info(sprintf('Found %d urls for pattern "%s"', count($patternUrls), $page));
-                $urls = array_merge($urls, $patternUrls);
-            } else if (strpos($page, 'http') === 0) {
-                if (!$this->isRelatedDomain($page)) {
-                    $this->logger->error(
-                        sprintf(
-                            'Page "%s" can\'t be warmed-up because such domain ' .
-                            'is not registered in current Magento installation',
-                            $page
-                        )
-                    );
-                } else {
-                    $urls[] = $page;
-                }
-            } else {
-                $urls[] = $baseUrl . '/' . $page;
-            }
-        }
-
-        return $urls;
-    }
-
-    /**
-     * Checks that host from $url is using in current Magento installation
-     *
-     * @param string $url
-     * @return bool
-     */
-    private function isRelatedDomain(string $url): bool
-    {
-        if ($this->baseHosts === null) {
-            $this->baseHosts = array_map(
-                function ($baseHostUrl) {
-                    return parse_url($baseHostUrl, PHP_URL_HOST);
-                },
-                $this->urlManager->getBaseUrls()
-            );
-        }
-
-        return in_array(parse_url($url, PHP_URL_HOST), $this->baseHosts);
-    }
-
-    /**
-     * @param string $warmUpPattern
-     * @return array
-     */
-    private function getUrlsByPattern(string $warmUpPattern): array
-    {
-        try {
-            list($entity, $storeId, $pattern) = explode(':', $warmUpPattern);
-
-            $command = sprintf('config:get:rewrite-urls --entity-type="%s"', $entity);
-            if ($storeId && $storeId !== '*') {
-                $command .= sprintf(' --store_id="%s"', $storeId);
-            }
-
-            $result = $this->shell->execute($command);
-
-            $urls = json_decode($result[0]);
-
-            if (json_last_error() !== JSON_ERROR_NONE) {
-                $this->logger->error('Can\'t parse result from config:get:rewrite-urls: ' . json_last_error_msg());
-                return [];
-            }
-
-            if ($pattern === '*') {
-                return $urls;
-            }
-
-            $pattern = '/' . preg_quote($pattern) . '/';
-            $urls = array_filter($urls, function($url) use ($pattern) {
-                return preg_match($pattern, $url);
-            });
-
-            return $urls;
-
-        } catch (ShellException $e) {
-            $this->logger->error('Can\'t get result from config:get:rewrite-urls: ' . $e->getMessage());
-            return [];
-        }
-
     }
 }
